+++ src/main/data/godbot.txt
@@ -1,0 +1,3 @@
+T | 0 | read book
+D | 0 | submit ass | 2024-12-02
+E | 0 | die | |12pm |  12am
