--- src/main/java/TaskList.java
@@ -1,44 +1,0 @@
-import java.util.ArrayList;
-
-public class TaskList {
-    private ArrayList<Task> tasks;
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public TaskList(ArrayList<Task> loadedTasks) {
-        this.tasks = loadedTasks;
-    }
-
-    public void addTask(Task task) {
-        tasks.add(task);
-    }
-
-    public void deleteTask(int index) {
-        tasks.remove(index);
-    }
-
-    public Task getTask(int index) {
-        return tasks.get(index);
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-
-    public void showTasks(Ui ui) {
-        if (tasks.isEmpty()) {
-            ui.showMessage("No tasks for you mortal.");
-            return;
-        }
-        for (int i = 0; i < tasks.size(); i++) {
-            ui.showMessage((i + 1) + ". " + tasks.get(i));
-        }
-    }
-
-    public ArrayList<Task> getAllTasks() {
-        return tasks;
-    }
-}
-
