--- src/main/java/data/godbot.txt
@@ -1,3 +1,0 @@
-D | 0 | return book | 2023-12-10
-T | 0 | test
-D | 0 | submit report | 2024-02-10
