+++ src/main/java/godbot/exception/GodBotException.java
@@ -1,3 +1,5 @@
+package godbot.exception;
+
 public class GodBotException extends Exception {
   public GodBotException(String message) {
     super(message);
