+++ src/main/java/godbot/task/Event.java
@@ -1,3 +1,4 @@
+package godbot.task;
 public class Event extends Task{
     private String from;
     private String to;
