+++ src/main/java/godbot/task/Task.java
@@ -1,3 +1,5 @@
+package godbot.task;
+
 public abstract class Task {
   protected String description;
   protected boolean isDone;
