+++ src/main/java/godbot/task/ToDo.java
@@ -1,3 +1,5 @@
+package godbot.task;
+
 public class ToDo extends Task{
 
     public ToDo(String description, boolean isDone){
