+++ src/main/java/godbot/ui/Ui.java
@@ -1,3 +1,4 @@
+package godbot.ui;
 import java.util.Scanner;
 
 public class Ui {
